third_party/WebKit/LayoutTests/http/tests/inspector/network/network-worker-fetch-expected.txt

Tests fetch in worker.

Fetch in worker result: Hello world
http://127.0.0.1:8000/inspector/network/resources/fetch-worker.js
resource.type: other
request.failed: false
http://127.0.0.1:8000/inspector/network/resources/resource.php
resource.type: other
request.failed: false
